Change many int/long to Py_ssize_t/size_t/NyBit
C long is 32 bit even on 64 bit Windows... Why does MS just have to
be different, every time?

+++ src/heapy/heapdef.h
@@ -62,7 +62,7 @@ typedef struct NyHeapRelate {
 
 /* Definitions of its function types, useful for casting. */
 
-typedef int (*NyHeapDef_SizeGetter) (PyObject *obj);
+typedef size_t (*NyHeapDef_SizeGetter) (PyObject *obj);
 typedef int (*NyHeapDef_Traverser) (NyHeapTraverse *arg);
 typedef int (*NyHeapDef_RelationGetter) (NyHeapRelate *r);
